app/requirements.txt
Flask>=2.3
pydantic>=2.0
mip
injector
python-ulid
ortools>=9.0